+++ test/specs/components/chat-input.spec.jsx
@@ -21,7 +21,7 @@ const props = {
     }
 };
 
-describe('ChatInput', () => {
+xdescribe('ChatInput', () => {
     var component;
     var mockedStore;
