+++ test/specs/components/settings.spec.jsx
@@ -2,204 +2,50 @@ import sinon from 'sinon';
 import React from 'react';
 import TestUtils from 'react-addons-test-utils';
 
-import { SettingsComponent } from '../../../src/js/components/settings';
-import * as userService from '../../../src/js/services/user-service';
+import { Settings } from '../../../src/js/components/settings';
+import { NotificationsSettings } from '../../../src/js/components/notifications-settings';
+import { EmailSettings } from '../../../src/js/components/email-settings';
+import * as appUtils from '../../../src/js/utils/app';
+import { mockComponent } from '../../utils/react';
+import { createMockedStore } from '../../utils/redux';
+
+import { ParentComponentWithContext } from '../../utils/parent-component';
+
 
 const sandbox = sinon.sandbox.create();
 const defaultProps = {
-    appState: {
-        settingsNotificationVisible: false,
-        readOnlyEmail: false
-    },
-    ui: {
-        text: {
-            settingsReadOnlyText: 'This is readonly',
-            settingsText: 'This is settings',
-            settingsInputPlaceholder: 'This is a placeholder',
-            settingsSaveButtonText: 'This is a button text'
-        }
-    },
-    user: {
-        email: 'some@email.com'
-    }
+    className: 'class-name'
 };
 
 
 describe('Settings', () => {
 
-    var component;
-
     beforeEach(() => {
-        sandbox.stub(userService, 'immediateUpdate');
-        userService.immediateUpdate.resolves();
+        sandbox.stub(appUtils, 'hasChannels');
+        appUtils.hasChannels.resolves(false);
+
+        mockComponent(sandbox, NotificationsSettings, 'div', {
+            className: 'notificationSettings'
+        });
+        mockComponent(sandbox, EmailSettings, 'div', {
+            className: 'emailSettings'
+        });
     });
 
     afterEach(() => {
         sandbox.restore();
     });
 
-    describe('Email read-only', () => {
-        var props = Object.assign({}, defaultProps, {
-            appState: Object.assign({}, defaultProps.appState, {
-                readOnlyEmail: true
-            })
-        });
+    it('Should render the NotificationSettings component', () => {
 
-        beforeEach(() => {
-            component = TestUtils.renderIntoDocument(<SettingsComponent {...props} />);
-        });
-
-        it('should render the read-only text', () => {
-            component.refs.description.textContent.should.eq(props.ui.text.settingsReadOnlyText);
-        });
-
-        it('should put user email in input', () => {
-            component.refs.input.value.should.eq(props.user.email);
-        });
-
-        it('should disable the input', () => {
-            component.refs.input.disabled.should.be.true;
-        });
-
-        it('should not have errors', () => {
-            expect(component.refs.button).to.not.exist;
-            TestUtils.scryRenderedDOMComponentsWithClass(component, 'has-error').length.should.be.eq(0);
-        });
+        const props = Object.assign({}, defaultProps, {});
+        const store = createMockedStore(sandbox, props);
+        const context = {settings: {}};
+        const component = TestUtils.renderIntoDocument(<ParentComponentWithContext context={ context } store={ store }>
+                <Settings {...props} />
+            </ParentComponentWithContext>);
+        TestUtils.scryRenderedDOMComponentsWithClass(component, 'notificationSettings').length.should.eq(1);
+        TestUtils.scryRenderedDOMComponentsWithClass(component, 'emailSettings').length.should.eq(0);
     });
 
-    describe('Email editable', () => {
-        var props = Object.assign({}, defaultProps);
-
-        beforeEach(() => {
-            component = TestUtils.renderIntoDocument(<SettingsComponent {...props} />);
-        });
-
-        it('should render the normal text', () => {
-            component.refs.description.textContent.should.eq(props.ui.text.settingsText);
-        });
-
-        it('should put user email in input', () => {
-            component.refs.input.value.should.eq(props.user.email);
-        });
-
-        it('should enable the input', () => {
-            component.refs.input.disabled.should.be.false;
-        });
-
-        it('should not have errors', () => {
-            component.refs.button.disabled.should.be.false;
-            TestUtils.scryRenderedDOMComponentsWithClass(component, 'has-error').length.should.be.eq(0);
-        });
-    });
-
-    describe('Input', () => {
-        var props = Object.assign({}, defaultProps);
-
-        beforeEach(() => {
-            sandbox.stub(SettingsComponent.prototype, 'onChange');
-            component = TestUtils.renderIntoDocument(<SettingsComponent {...props} />);
-        });
-
-        it('should call onChange', () => {
-            TestUtils.Simulate.change(component.refs.input);
-            SettingsComponent.prototype.onChange.should.have.been.calledOnce;
-        });
-
-
-    });
-
-    describe('Save button', () => {
-        var props = Object.assign({}, defaultProps, {
-            user: {
-                email: 'some@email.com'
-            }
-        });
-
-        beforeEach(() => {
-            sandbox.stub(SettingsComponent.prototype, 'save');
-            component = TestUtils.renderIntoDocument(<SettingsComponent {...props} />);
-        });
-
-        it('should call save', () => {
-            TestUtils.Simulate.click(component.refs.button);
-            SettingsComponent.prototype.save.should.have.been.calledOnce;
-        });
-    });
-
-    describe('Save', () => {
-        var props = Object.assign({}, defaultProps, {
-            user: {
-                email: 'some@email.com'
-            }
-        });
-
-        var event;
-
-        beforeEach(() => {
-            Object.assign(props, {
-                actions: {
-                    hideSettings: sandbox.stub()
-                }
-            });
-
-            component = TestUtils.renderIntoDocument(<SettingsComponent {...props} />);
-            sandbox.spy(component, 'setState');
-            event = {
-                preventDefault: sandbox.stub()
-            };
-        });
-
-        it('should prevent button default behavior', () => {
-            return component.save(event).then(() => {
-                event.preventDefault.should.have.been.calledOnce;
-            });
-        });
-
-        describe('valid email', () => {
-            it('should call immediateUpdate and hideSettings', () => {
-                return component.save(event).then(() => {
-                    userService.immediateUpdate.should.have.been.calledOnce;
-                });
-
-            });
-        });
-
-        describe('invalid email', () => {
-
-            beforeEach(() => {
-                component.setState({
-                    email: 'invalid email value'
-                });
-            });
-
-
-            it('should update state with error', () => {
-                return component.save(event).then(() => {
-                    component.setState.should.have.been.calledWith({
-                        hasError: true
-                    });
-                });
-            });
-        });
-
-
-        describe('empty email', () => {
-
-            beforeEach(() => {
-                component.setState({
-                    email: ''
-                });
-            });
-
-
-            it('should update state with error', () => {
-                return component.save(event).then(() => {
-                    component.setState.should.have.been.calledWith({
-                        hasError: true
-                    });
-                });
-            });
-        });
-
-    });
 });
